+++ src/main/java/com/starkinfra/utils/Response.java
@@ -37,8 +37,7 @@ public final class Response {
 
     public String content() throws java.io.IOException {
         StringBuilder textBuilder = new StringBuilder();
-        try (Reader reader = new BufferedReader(new InputStreamReader
-                (stream, Charset.forName(StandardCharsets.UTF_8.name())))) {
+        try (Reader reader = new BufferedReader(new InputStreamReader (stream, Charset.forName(StandardCharsets.UTF_8.name())))) {
             int c;
             while ((c = reader.read()) != -1) {
                 textBuilder.append((char) c);
@@ -48,6 +47,9 @@ public final class Response {
     }
 
     public static Response fetch(String path, String method, JsonObject payload, Map<String, Object> query, User user) throws Exception {
+        return Response.fetch(path, method, payload, query, user, null, true);
+    }
+    public static Response fetch(String path, String method, JsonObject payload, Map<String, Object> query, User user, String prefix, Boolean raiseException) throws Exception {
         user = Check.user(user);
         String language = Check.language();
 
@@ -70,11 +72,16 @@ public final class Response {
         headers.put("Access-Id", user.accessId());
         headers.put("Access-Time", accessTime);
         headers.put("Access-Signature", Ecdsa.sign(message, user.privateKey()).toBase64());
-        headers.put("User-Agent", getUserAgent());
+        headers.put("User-Agent", getUserAgent(prefix));
         headers.put("Content-Type", "application/json");
         headers.put("Accept-Language", language);
 
         Response response = executeMethod(user, path, method, body, headers);
+
+        if (!raiseException) {
+            return response;
+        }
+
         if (response.status == 400) {
             throw new InputErrors(response.content());
         }
@@ -129,7 +136,8 @@ public final class Response {
         return new Response(status, contentStream);
     }
 
-    private static String getUserAgent() {
-        return (userAgentOverride == null) ? "Java-" + System.getProperty("java.version") + "-SDK-Infra-0.11.3" : userAgentOverride;
+    private static String getUserAgent(String prefix) {
+        prefix = prefix != null ? prefix += "-" : null;
+        return (userAgentOverride == null) ? prefix + "Java-" + System.getProperty("java.version") + "-SDK-Infra-0.11.3" : userAgentOverride;
     }
 }